chore : SelfModeUsecase와 Protocol 파일 분리
